Fix symlink test and add harlink test
